Add .mif Generator for Row RAM
I added a .mif generator script to help make a simple graphics
test for the Row RAM. The .mif files specify data, which can
optionally be loaded into RAMs during initialization. The test I
wrote is intended to test every color in the first palette.

The .gitignore had to be modified to ignore all .mif files.