Implement PPU FSM; Remove VRAM Swap
The PPU FSM has been finished. The PPU starts in a PPU_SYNC state,
waiting for the hdmi_video_output's display period to begin. When
the display period starts, the PPU sends an interrupt to the CPU.
When the display period ends, the PPU FSM goes into a PPU_SYNC or
PPU_LATE state depending on whether the CPU has finished writing.
In the PPU_SYNC state, the VRAMs are synchronized. If the
PPU_LATE state occurs, the CPU is given an extra display and
blank period to finish its data transfers to VRAM.

As it turns out, VRAM swap is not needed when the VRAMs synchronize
between themselves. I removed the VRAM Swap implementation in favor
of a fixed PPU-Facing VRAM and CPU-Facing VRAM, with interconnect
controlled by vram_interconnect.sv. This interconnect module
manages which of the PPU, CPU, or vram_sync_writer.sv module gets
access to the VRAM. The PPU FSM signals to the vram_interconnect
when to change the connections.

Small additions:
* SystemVerilog ModPorts were added to the vram_if interface to
make port-directions clearer.