Fix Row-Buffer Swap Timings
Renamed vblank_end to vblank_end_soon to reflect what the signal
actually means. Changed the time the signal is sent to the 33rd
(still in vblank) row. This is to give the PPU-logic the same
duration to prepare the row buffer as any other time in the
display period.

Corrected the VRAM Swap signal behavior. The signal should be sent
only on every other row starting on odd rows. This is the correct
behaviour: The PPU will start buffering the first row at row #33.
Then the row buffers are swapped on the first display row (#35),
that row is reused for row #36 and isn't swapped until the start
of row #37. This effectively upscales the image by x2 vertically.