Implement Tile Mirror X; Fix Pixel Mixer Fetching
Apparently I forgot to implement the horizontal tile mirror. I've
added it in, though the implementation wasn't as straight-forward
as I thought it would be. It required use of two buffers for the
x-mirror signal after accessing it in the local tile-RAM buffer,
so that it could be used again when saving the data fetched from
Pattern-RAM. This feature is tested using the ppu_logic_tb.sv
testbench.

Fixed an issue where Pixel-Mixer was ending the fetch/write
process at pixel 317 instead of pixel 319.