Fix Y-Mirror and X-Scroll
The solution to the Y-Mirror graphical glitches was to preserve
the y-mirror signal from the moment it is first read from the
tilram_rbuf until the patram_fetcher is ready to write pattern
data to-be-mirrored into patram_rbuf. Then, the we can reverse the
ordering of rows stored to achieve the mirroring effect.
Shout-out to Andrew for helping find where the issue was!

Fixed X-scrolling. The error was that not all 44 tiles stored in
tilram_rbuf were being fetched, and the logic for the pixel-mixer
to address a given pixel with scroll applied was incorrect.

Also, I've refactored tile_engine.sv a bit to make it easier to
read and find where interconnects between submodules belong.