Finalize Video Demo
Uploaded the VRAM converter scripts I used to generate the various
VRAM segment initializers (.hex files) used in the demo.

Increased the scrolling speed so that it isn't painfully slow.

Replaced ugly lines in the demo tile map with regular walk-way
tiles. They didn't line up at the boundaries of the screen, so
they stood out when scrolling past the 64x64 edges of the Tile-RAM.